Add files via upload
Learned a lot of new things in this project. Learned about the complexity of using dates and reading that data. Also learned about using steppers as a way for users to input data. An interesting part of this project was using CoreML and seeing how easy Apple makes it to use Machine Learning technology in apps.