Add files via upload
made some changes to the app and attempted to make some others. Attempted to add an edit button alongside the plus button, was unsuccessful for some reason. However, I was successful in adding an alert for the expense price. Now if someone tries to enter something other than a number in that field, they receive an error.